+++ tests/unit/validators/HexColorValidatorTest.php
@@ -26,9 +26,19 @@ class HexColorValidatorTest extends Unit
 
         $model->color = 'fff';
         self::assertTrue($model->validate());
+        self::assertEquals('fff', $model->color);
+
+        $model->color = '#fff';
+        self::assertTrue($model->validate());
+        self::assertEquals('fff', $model->color);
 
         $model->color = '000000';
         self::assertTrue($model->validate());
+        self::assertEquals('000000', $model->color);
+
+        $model->color = '#000000';
+        self::assertTrue($model->validate());
+        self::assertEquals('000000', $model->color);
 
         $model->color = 'invalid';
         self::assertFalse($model->validate());