+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.2.4 (under development)
 
+- Added mailer transport DSN fallback (Issues #24) 
 - Enhanced `Nav` to allow for `active` callables
 - Enhanced `Navbar` item sorting by introducing an optional `order` array key
 - Enhanced `SignupForm` to better handle the token validation