+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.4 (under development)
 
 - Added mailer transport DSN fallback (Issues #24) 
+- Added support for `color`  type input fields (Issue #22)
 - Enhanced `Nav` to allow for `active` callables
 - Enhanced `Navbar` item sorting by introducing an optional `order` array key
 - Enhanced `SignupForm` to better handle the token validation